Including Foreign Key of monitored objects in analysis, alert uses dataset Foreign Key to identify monitored objects id, alter uses of [] to at in maps to avoid creating invalid objects, temporarily comment additional values in alert - not working

+++ src/examples/core/AccessLastDcpInpeFromLocal.cpp
@@ -49,7 +49,7 @@ int main(int argc, char* argv[])
 
     //empty filter
     terrama2::core::Filter filter;
-    filter.lastValues = std::make_shared<int32_t>(1);
+    filter.lastValues = std::make_shared<size_t>(1);
 
     //accessing data
     terrama2::core::DataAccessorDcpInpe accessor(dataProviderPtr, dataSeriesPtr);
